Add initial location search
This location search component uses the Google Places Autocomplete API
to provide a list of locations to choose from. This helps ensure we get
a valid location from the user which includes a City, State, Postal
Code, latitude, and longitude.

This also allows us to forego additional geocoding requests to Google.